+++ src/components/EurobonusDashboard.tsx
@@ -6,10 +6,12 @@ import type { Transaction } from "../models/transaction";
 import type { Profile } from "../models/profile";
 import { Spinner } from "@blueprintjs/core";
 import { calculateTotalEuroBonusPoints, getEarliestDate, getLatestDate } from "../utils/calculations";
+import { useSummarizedPoints } from "../hooks/useSummarizedPoints";
 
 export const EurobonusDashboard = () => {
   const { data, loading } = useFetchTransactions();
   const { data: profilesData, loading: profileLoading } = useFetchProfiles();
+  const { data: summarizedPoints, loading: loadingSummarized } = useSummarizedPoints();
 
   const transactions: Transaction[] = useMemo(() => (data ? data : []), [data]);
   const profiles: Profile[] = useMemo(() => (profilesData ? profilesData : []), [profilesData]);
@@ -17,7 +19,7 @@ export const EurobonusDashboard = () => {
   const earliestdate = useMemo(() => getEarliestDate(transactions), [transactions]);
   const latestDate = useMemo(() => getLatestDate(transactions), [transactions]);
 
-  if (loading || profileLoading) {
+  if (loading || profileLoading || loadingSummarized) {
     return <Spinner />;
   }
   if (transactions.length === 0) {