+++ src/components/TotalDashboard.tsx
@@ -4,10 +4,12 @@ import type { Transaction } from "../models/transaction";
 import type { Profile } from "../models/profile";
 import { Spinner } from "@blueprintjs/core";
 import useFetchTotalTransactions from "../hooks/useFetchTotalTransactions";
-import { calculateTotalEuroBonusPoints } from "../utils/calculations";
+import { calculateAveragePoints, calculateTotalEuroBonusPoints } from "../utils/calculations";
+import { useSummarizedPoints } from "../hooks/useSummarizedPoints";
 
 export const TotalDashboard = () => {
   const { data, loading } = useFetchTotalTransactions();
+  const { data: summarizedPoints, error } = useSummarizedPoints();
 
   const profile: Profile = useMemo(() => ({ id: "1", display_name: "All Users", created: new Date(2025, 1, 20).toISOString(), user_id: "1" }), []);
   const transactions: Transaction[] = useMemo(
@@ -20,9 +22,21 @@ export const TotalDashboard = () => {
   if (loading) {
     return <Spinner />;
   }
+
+  if (error) {
+    return <p>Error: {error}</p>;
+  }
+
+  const average = calculateAveragePoints(summarizedPoints);
+
   return (
     <>
-      {sum > 0 && <p>{`Users have earned a total of ${sum.toLocaleString()} eurobonus points since 2017`}</p>}
+      {sum > 0 && (
+        <div style={{ display: "flex", justifyContent: "space-between" }}>
+          <p> {`Users have earned a total of ${sum.toLocaleString()} eurobonus points since 2017`}</p>
+          <p>{`Average per user: ${average.toLocaleString()}`}</p>
+        </div>
+      )}
       <EurobonusCharts transactions={transactions} profiles={[profile]} hideTable />
     </>
   );